Disable imagelist use below Vista

+++ Core/KanbanBoard/KanbanCtrl.cpp
@@ -3770,8 +3770,11 @@ void CKanbanCtrl::OnCaptureChanged(CWnd* pWnd)
 
 	m_aColumns.SetDropTarget(NULL);
 
-	m_ilDrag.EndDrag();
-	m_ilDrag.DeleteImageList();
+	if (m_ilDrag.GetSafeHandle())
+	{
+		m_ilDrag.EndDrag();
+		m_ilDrag.DeleteImageList();
+	}
 }
 
 BOOL CKanbanCtrl::CanDragTask(DWORD dwTaskID) const
@@ -3948,8 +3951,11 @@ void CKanbanCtrl::OnMouseMove(UINT nFlags, CPoint point)
 	if (IsDragging())
 	{
 		// get the list and item under the mouse
-		ClientToScreen(&point);
-		m_ilDrag.DragMove(point);
+		if (m_ilDrag.GetSafeHandle())
+		{
+			ClientToScreen(&point);
+			m_ilDrag.DragMove(point);
+		}
 
 		const CKanbanColumnCtrl* pDestCol = m_aColumns.HitTest(point);
 		BOOL bValidDest = CanEndDrag(m_pSelectedColumn, pDestCol);
